refactor: apis 타입스크립트 전환
