--- src/apis/client.js
@@ -1,9 +1,0 @@
-import axios from 'axios';
-
-const baseURL = process.env.REACT_APP_HOST;
-
-const client = axios.create({
-  baseURL,
-});
-
-export default client;
